Fix default indent_width
Even though it's configurable in style.py, indent_width was stuck to 2 due to
a param of FormatCode. Fix this and add a test that a different indent width
works as well. Command-line tests were modified to make sure that the current
default (4) is actually used when style.py is not mucked with.