Add a default value to GetNodeAnnotation
Some node annotations are known to be numeric. For those, it makes sense to have
a default value of 0 when querying, because the result of GetNodeAnnotation is
directly compared with other integers.

Add a default setting and use it everywhere that's relevant.

This lets us pass some reformatting tests on Python 3!