Replace print statement by a print() call.
Using print statements is problematic in formatted code because Python 3
fails in VerifyCode. print() is accepted by both.

Not sure how critical it is to be able to test Python2-style print statements.
If it is, we may want to rig VerifyCode to do a version distinction.

This change lets us pass all reformatter tests on Python 3!