Opens the README.rst file with an explicit encoding.
This is important because of the difference between bytes and strings
and the follies of `sys.getdefaultencoding()` in guessing the
appropriate encoding. Be explicit; use the utf-8 encoding for the
README.rst file, and decode it explicitly as utf-8-encoded text.

Uses the `codecs.open` function for cross-compatibility between Python 2
and Python 3. In Python 3, `open` takes a keyword argument of
`encoding`, but unfortunately `open` in Python 2 does not. `codecs.open`
solves this issue by taking an explicit encoding as an argument and by
being available in the standard library in both Python 2 and Python 3.