Add support for project-local style (in .style.yapf)
YAPF will now look for this file at startup and use it as
the default if found.  If not found pep8 is still the default.

Now that we have project-local style there is no need for the
hardcoded yapf style (it will be the default for all files
in the yapf tree now).