Open files with an encoding.
If we don't open files with the correct encoding, it will lose information when
reading and writing the files. In particular, if there's a byte order mark
(BOM), then it will not detect it, which causes an error to occur in the
parsing. And then when we write the file, that BOM is lost.

Closes #50