Make return None instead of '' for empty diffs
FormatFiles in __init__.py tests if the result of yapf_api.FormatFile
is None before outputting it. However, it returned '' if print_diff
was True and the reformatted code was the same as the input code, i.e.
it returned '' when there was no diff instead of None. This changes
yapf_api.FormatFile to return None in this case.