Stablize reformatting with Facebook style
The line number cannot be used to determine if the line should be split when
"dedenting" is specified. The reason for this is because a newline may be added
without splitting around the brackets. Then if we rerun yapf on the output, the
code will be on different lines, and therefore be formatted again.

This change doesn't rely upon the token's original line number.

Closes #210