Determine if the opening bracket was split
This heuristic is only partially correct. It uses the opening bracket's column
in the unformatted source to determine if it was split when looking at the
closing bracket. This is because there's no way in yapf to determine if a token
is going to be split in the final output until the final line is reconstructed.