Split after an opening bracket if named assigns
When we intentionally split each argument because of a named assignment, then
also split after the opening bracket (before the first argument). This makes
things a bit nicer to look at. It also allows one to rename the function
without having to reformat all of the arguments to the function call.