Update PEP 8 style to match modern PEP 8
PEP 8 was updated last year to prefer that line splits come before binary operators, rather than after:

https://www.python.org/dev/peps/pep-0008/#should-a-line-break-before-or-after-a-binary-operator
https://github.com/python/peps/commit/c59c4376ad233a62ca4b3a6060c81368bd21e85b

Really there should be a way to do this for other binary operations... for example, even with this setting, YAPF wants to do:

```
     orig_extracted = (
-        extract_tb(orig.__traceback__)
-        + extract_tb(orig.exceptions[0].__traceback__)
-        + extract_tb(orig.exceptions[0].exceptions[1].__traceback__))
+        extract_tb(orig.__traceback__) +
+        extract_tb(orig.exceptions[0].__traceback__) +
+        extract_tb(orig.exceptions[0].exceptions[1].__traceback__))
```

which is exactly what the example in PEP 8 says not to do. But this patch brings YAPF's PEP 8 style closer to what PEP 8 says.