Treat 'async def' like 'def'
yapf has special treatment for (a) compound statements, and (b)
compound statements that are not 'def' statements. So you would expect
that 'async def' would be in category (a), but not in category (b).

Originally, 'async def' was not in either category, but this was fixed
in 6412ce704a8f, which made 'async def' (and other async constructs)
treated like compound statements. However, the special check for 'def'
statements was *not* extended to 'async def', so now 'async def' was
in categories (a) *and* (b).

This commit fixes that, so that 'async def' is in category (a) but not
category (b).

Fixes gh-458