Fix split penalty for single-argument function calls with generators.
The previous logic considered all components of the generator in a
single-argument call such as 'f(a for a in x)' to be strongly connected
like a normal single-argument call (e.g. 'f(a_var)'). This caused
formatting irregularities in these cases.

This change retains normal split penalties in the single-argument
generator case.