Test if a string is a multiline string without using a regex.
The approach of using a regex makes sure to exclude any token that isn't a multiline string by only matching valid multiline strings. But at this point all tokens are already known to be valid, so a faster heuristic can be used.

Instead by taking the approach of checking to see if the last 3 characters are a triple quotation mark, testing for a multiline string is significantly faster.