Allow `lines` parameter to fix whitespace between statements. (#864)
* Fix split penalty for single-argument function calls with generators.

The previous logic considered all components of the generator in a
single-argument call such as 'f(a for a in x)' to be strongly connected
like a normal single-argument call (e.g. 'f(a_var)'). This caused
formatting irregularities in these cases.

This change retains normal split penalties in the single-argument
generator case.

* Allow `lines` parameter to fix whitespace between statements.

The previous logic largely ignored `lines` values that were not part of
statements. This means that yapf required a statement to be formatted to
make whitespace-only changes.

This logic tweak allows `lines` to be taken into account even when the
adjacent statements are disabled (i.e. not in `lines`).

* Remove chromium style from test.